requirements.txt
aiohttp==3.13.2
aiomysql[rsa]==0.3.2
beautifulsoup4[lxml]==4.14.3
discord.py==2.6.4
google-api-python-client==2.187.0
isodate==0.7.2
python-dotenv==1.2.1
